Refactor pause point status collection for repository limits
Split status-list parsing from the await path and move status collection out of the registry so the Go and C# implementations remain below the enforced complexity and file-length limits without changing status behavior.
