+++ Packages/src/Runtime/PausePoints/UloopPausePointStatusSnapshotCollector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85bd573987fde4673b4b783ef96658df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
