+++ cli/project-runner/internal/projectrunner/pause_point_status_list.go
@@ -1,0 +1,40 @@
+package projectrunner
+
+import (
+	"context"
+	"encoding/json"
+	"io"
+
+	clierrors "github.com/hatayama/unity-cli-loop/common/errors"
+
+	"github.com/hatayama/unity-cli-loop/common/clicore"
+	"github.com/hatayama/unity-cli-loop/common/unityipc"
+)
+
+func runPausePointStatusListCommand(
+	ctx context.Context,
+	connection unityipc.Connection,
+	stdout io.Writer,
+	stderr io.Writer,
+) int {
+	response, err := queryPausePointStatusList(ctx, connection)
+	if err != nil {
+		clierrors.WriteClassifiedError(stderr, err, clierrors.ErrorContext{
+			ProjectRoot: connection.ProjectRoot,
+			Command:     clicore.PausePointStatusUserCommandName,
+		})
+		return 1
+	}
+
+	result, err := json.Marshal(response)
+	if err != nil {
+		clierrors.WriteClassifiedError(stderr, err, clierrors.ErrorContext{
+			ProjectRoot: connection.ProjectRoot,
+			Command:     clicore.PausePointStatusUserCommandName,
+		})
+		return 1
+	}
+
+	clicore.WriteJSON(stdout, result)
+	return 0
+}
